Create createAsyncThunk to ingredients
