+++ src/redux/ingredients-slice.ts
@@ -1,11 +1,14 @@
-import { createSlice, PayloadAction } from '@reduxjs/toolkit';
+import { createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit';
+import { getIngredientData } from '../api/agent';
 import { bunName } from '../common/constants';
 import { IIngredientsState } from '../models/app-store';
 import { IIngredientsItem } from '../models/ingredients';
+import { setAppError } from './app-slice';
 import { addOrDeleteCountIngredients } from './common';
 import {
   addIngredientInConstructor,
   deleteIngredientFromConstructor,
+  setBun,
 } from './constructor-ingredients-slice';
 
 const initialState: IIngredientsState = {
@@ -13,6 +16,24 @@ const initialState: IIngredientsState = {
   countIngredients: [],
   isLoading: false,
 };
+
+export const ingredientFetch = createAsyncThunk(
+  'ingredients/ingredientFetch',
+  async function (_, { rejectWithValue, dispatch }) {
+    try {
+      const ingredientsData = await getIngredientData();
+      const firstBunElement = ingredientsData.filter(
+        (item) => item.type === bunName
+      )[0];
+
+      dispatch(setBun(firstBunElement));
+      dispatch(setIngredients(ingredientsData));
+    } catch (error) {
+      rejectWithValue(error);
+      dispatch(setAppError(error));
+    }
+  }
+);
 
 const ingredientsSlice = createSlice({
   name: 'ingredients',
@@ -26,6 +47,18 @@ const ingredientsSlice = createSlice({
     },
   },
   extraReducers: {
+    // @ts-expect-error
+    [ingredientFetch.pending]: (state) => {
+      state.isLoading = true;
+    },
+    // @ts-expect-error
+    [ingredientFetch.fulfilled]: (state) => {
+      state.isLoading = false;
+    },
+    // @ts-expect-error
+    [ingredientFetch.rejected]: (state) => {
+      state.isLoading = false;
+    },
     // @ts-expect-error
     [addIngredientInConstructor]: (
       state,
